+++ ingenialink/__init__.py
@@ -54,4 +54,4 @@ __all__ = [
     "CanopenDictionary",
 ]
 
-__version__ = "7.1.0"
+__version__ = "7.1.1"