+++ CHANGELOG.md
@@ -1,9 +1,20 @@
 # Changelog
 
-## [7.2.0]
+## [Unreleased]
+
+### Add
+- Motor enable and disable features in the virtual drive.
+- Emulate control loops in the virtual drive.
+- EtherCAT PDO module.
+
+
+## [7.1.1] - 2024-01-03
 
 ### Added
-- EtherCAT PDO module.
+- Missing EtherCAT protocol documentation.
+
+### Fixed
+- Import Ingenialink does not raise an error if WinPcap is not installed, but ethercat features are disabled.
 
 ## [7.1.0] - 2023-11-28